move 'MessageWaitedType' and 'WasmEntryPoint' to 'gear-core/env' module
