+++ core-backend/src/error.rs
@@ -19,9 +19,9 @@
 use actor_system_error::actor_system_error;
 use gear_core::{
     buffer::PanicBuffer,
+    env::MessageWaitedType,
     gas::{ChargeError, CounterType},
     ids::ProgramId,
-    message::MessageWaitedType,
 };
 use gear_core_errors::ExtError as FallibleExtError;
 use parity_scale_codec::{Decode, Encode};
